LayoutTests/platform/mac-monterey-wk1/imported/w3c/web-platform-tests/cookies/value/value-ctl-expected.txt

CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/cookies/resources/cookie.py?set=%22test0%5Cu0000name%3D0%22.

FAIL Cookie with %x0 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=0"
FAIL Cookie with %x1 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=1"
FAIL Cookie with %x2 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=2"
FAIL Cookie with %x3 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=3"
FAIL Cookie with %x4 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=4"
FAIL Cookie with %x5 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=5"
FAIL Cookie with %x6 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=6"
FAIL Cookie with %x7 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=7"
FAIL Cookie with %x8 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=8"
PASS Cookie with %x9 in value is accepted (DOM).
FAIL Cookie with %xa in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=10"
FAIL Cookie with %xb in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=11"
FAIL Cookie with %xc in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=12"
FAIL Cookie with %xd in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=13"
FAIL Cookie with %xe in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=14"
FAIL Cookie with %xf in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=15"
FAIL Cookie with %x10 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=16"
FAIL Cookie with %x11 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=17"
FAIL Cookie with %x12 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=18"
FAIL Cookie with %x13 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=19"
FAIL Cookie with %x14 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=20"
FAIL Cookie with %x15 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=21"
FAIL Cookie with %x16 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=22"
FAIL Cookie with %x17 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=23"
FAIL Cookie with %x18 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=24"
FAIL Cookie with %x19 in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=25"
FAIL Cookie with %x1a in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=26"
FAIL Cookie with %x1b in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=27"
FAIL Cookie with %x1c in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=28"
FAIL Cookie with %x1d in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=29"
FAIL Cookie with %x1e in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=30"
FAIL Cookie with %x1f in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=31"
FAIL Cookie with %x7f in value is rejected (DOM). assert_equals: The cookie was rejected. expected "" but got "test=127"
PASS Cookie with %x0 in name is rejected or modified (HTTP).
FAIL Cookie with %x1 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=1"
FAIL Cookie with %x2 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=2"
FAIL Cookie with %x3 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=3"
FAIL Cookie with %x4 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=4"
FAIL Cookie with %x5 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=5"
FAIL Cookie with %x6 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=6"
FAIL Cookie with %x7 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=7"
FAIL Cookie with %x8 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=8"
PASS Cookie with %x9 in value is accepted (HTTP).
FAIL Cookie with %xa in name is rejected or modified (HTTP). assert_equals: The cookie was set as expected. expected "test10 name=10" but got ""
FAIL Cookie with %xb in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=11"
FAIL Cookie with %xc in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=12"
FAIL Cookie with %xd in name is rejected or modified (HTTP). assert_equals: The cookie was set as expected. expected "test13 name=13" but got ""
FAIL Cookie with %xe in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=14"
FAIL Cookie with %xf in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=15"
FAIL Cookie with %x10 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=16"
FAIL Cookie with %x11 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=17"
FAIL Cookie with %x12 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=18"
FAIL Cookie with %x13 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=19"
FAIL Cookie with %x14 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=20"
FAIL Cookie with %x15 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=21"
FAIL Cookie with %x16 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=22"
FAIL Cookie with %x17 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=23"
FAIL Cookie with %x18 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=24"
FAIL Cookie with %x19 in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=25"
FAIL Cookie with %x1a in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=26"
FAIL Cookie with %x1b in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=27"
FAIL Cookie with %x1c in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=28"
FAIL Cookie with %x1d in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=29"
FAIL Cookie with %x1e in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=30"
FAIL Cookie with %x1f in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=31"
FAIL Cookie with %x7f in value is rejected (HTTP). assert_equals: The cookie was rejected. expected "" but got "test=127"
